requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.30.1
pydantic==2.7.1
pydantic_core==2.18.2
requests==2.31.0
setuptools==69.5.1
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
colorama
deepgram-sdk
groq
python-dotenv
colorama
requests
keyboard
fastapi
uvicorn
numpy 
sounddevice 
cartesia
soundfile
websockets
PyAudio==0.2.14
pydub